LPS-134986 move DynamicInheritanceRoleSetContributor and RoleSetContributorContext to portal-search-spi
